refactor: reorganize admin routes and improve code structure
- Rename /admin/calculations → /admin/matches for clarity (reflects "Match Management")
- Rename /admin/roles → /admin/users for consistency (reflects "User Management")
- Fix Svelte component rendering in admin layout (use direct icon components instead of svelte:component)
- Improve code formatting and consistency across admin pages
